my_app/pyproject.toml
[project]
name = "my-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "guardrails-ai>=0.7.2",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-ollama>=1.0.1",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-postgres>=3.0.4",
    "numpy<2.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
